+++ lib/csv_plus_plus/cli_flag.rb
@@ -11,13 +11,8 @@ module CSVPlusPlus
   class CLIFlag
     extend ::T::Sig
 
-    # public_constant :SUPPORTED_CSVPP_FLAGS
-
     sig { returns(::String) }
     attr_reader :description
-
-    sig { returns(::T.proc.params(options: ::CSVPlusPlus::Options, v: ::String).void) }
-    attr_reader :handler
 
     sig { returns(::String) }
     attr_reader :long_flag
@@ -25,92 +20,70 @@ module CSVPlusPlus
     sig { returns(::String) }
     attr_reader :short_flag
 
-    sig do
-      params(
-        short_flag: ::String,
-        long_flag: ::String,
-        description: ::String,
-        handler: ::T.proc.params(options: ::CSVPlusPlus::Options, v: ::String).void
-      ).void
-    end
+    sig { params(short_flag: ::String, long_flag: ::String, description: ::String).void }
     # @param short_flag [String] A definition of the short/single-character flag
     # @param long_flag [String] A definition of the long/word-based flag
     # @param description [String] A description of what the flag does
-    # @param handler [Proc(Options, String)] A proc which is called to handle when this flag is seen
-    def initialize(short_flag, long_flag, description, handler)
+    def initialize(short_flag, long_flag, description)
       @short_flag = short_flag
       @long_flag = long_flag
       @description = description
-      @handler = handler
     end
   end
 
+  FLAG_HANDLERS = ::T.let(
+    {
+      backup: ->(options, _v) { options.backup = true },
+      create: ->(options, _v) { options.create_if_not_exists = true },
+      'key-values': lambda { |options, v|
+                      options.key_values =
+                        begin
+                          [v.split('=')].to_h
+                        rescue ::StandardError
+                          {}
+                        end
+                    },
+      'offset-columns': ->(options, v) { options.offset[0] = v },
+      'offset-rows': ->(options, v) { options.offset[1] = v },
+      output: ->(options, v) { options.output_filename = ::Pathname.new(v) },
+      verbose: ->(options, _v) { options.verbose = true }
+    },
+    ::T::Hash[::Symbol, ::T.proc.params(options: ::CSVPlusPlus::Options::Options, v: ::String).void]
+  )
+  public_constant :FLAG_HANDLERS
+
   SUPPORTED_CSVPP_FLAGS = ::T.let(
     [
-      ::CSVPlusPlus::CLIFlag.new(
-        '-b',
-        '--backup',
-        'Create a backup of the spreadsheet before applying changes.',
-        ->(options, _v) { options.backup = true }
-      ),
+      ::CSVPlusPlus::CLIFlag.new('-b', '--backup', 'Create a backup of the spreadsheet before applying changes.'),
       ::CSVPlusPlus::CLIFlag.new(
         '-c',
         '--create',
-        "Create the sheet if it doesn't exist.  It will use --sheet-name if specified",
-        ->(options, _v) { options.create_if_not_exists = true }
+        "Create the sheet if it doesn't exist.  It will use --sheet-name if specified"
       ),
       ::CSVPlusPlus::CLIFlag.new(
         '-g SHEET_ID',
         '--google-sheet-id SHEET_ID',
         'The id of the sheet - you can extract this from the URL: ' \
-        'https://docs.google.com/spreadsheets/d/< ... SHEET_ID ... >/edit#gid=0',
-        ->(options, v) { options.google_sheet_id = v }
+        'https://docs.google.com/spreadsheets/d/< ... SHEET_ID ... >/edit#gid=0'
       ),
       ::CSVPlusPlus::CLIFlag.new(
         '-k',
         '--key-values KEY_VALUES',
-        'A comma-separated list of key=values which will be made available to the template',
-        lambda do |options, v|
-          options.key_values =
-            begin
-              [v.split('=')].to_h
-            rescue ::StandardError
-              {}
-            end
-        end
+        'A comma-separated list of key=values which will be made available to the template'
       ),
       ::CSVPlusPlus::CLIFlag.new(
         '-n SHEET_NAME',
         '--sheet-name SHEET_NAME',
-        'The name of the sheet to apply the template to',
-        ->(options, v) { options.sheet_name = v }
+        'The name of the sheet to apply the template to'
       ),
       ::CSVPlusPlus::CLIFlag.new(
         '-o OUTPUT_FILE',
         '--output OUTPUT_FILE',
-        'The file to write to (must be .csv, .ods, .xls)',
-        ->(options, v) { options.output_filename = ::Pathname.new(v) }
+        'The file to write to (must be .csv, .ods, .xls)'
       ),
-      ::CSVPlusPlus::CLIFlag.new(
-        '-v',
-        '--verbose',
-        'Enable verbose output',
-        lambda { |options, _v|
-          options.verbose = true
-        }
-      ),
-      ::CSVPlusPlus::CLIFlag.new(
-        '-x OFFSET',
-        '--offset-columns OFFSET',
-        'Apply the template offset by OFFSET cells',
-        ->(options, v) { options.offset[0] = v }
-      ),
-      ::CSVPlusPlus::CLIFlag.new(
-        '-y OFFSET',
-        '--offset-rows OFFSET',
-        'Apply the template offset by OFFSET rows',
-        ->(options, v) { options.offset[1] = v }
-      )
+      ::CSVPlusPlus::CLIFlag.new('-v', '--verbose', 'Enable verbose output'),
+      ::CSVPlusPlus::CLIFlag.new('-x OFFSET', '--offset-columns OFFSET', 'Apply the template offset by OFFSET cells'),
+      ::CSVPlusPlus::CLIFlag.new('-y OFFSET', '--offset-rows OFFSET', 'Apply the template offset by OFFSET rows')
     ].freeze,
     ::T::Array[::CSVPlusPlus::CLIFlag]
   )