+++ lib/csv_plus_plus/writer/excel.rb
@@ -7,9 +7,18 @@ require_relative './rubyxl_builder'
 module CSVPlusPlus
   module Writer
     # A class that can output a +Template+ to an Excel file
-    class Excel < ::CSVPlusPlus::Writer::BaseWriter
+    class Excel < ::CSVPlusPlus::Writer::Writer
       extend ::T::Sig
       include ::CSVPlusPlus::Writer::FileBackerUpper
+
+      sig { params(options: ::CSVPlusPlus::Options::FileOptions, position: ::CSVPlusPlus::Runtime::Position).void }
+      # @param options [Options::FileOptions]
+      # @param position [Runtime::Position]
+      def initialize(options, position)
+        super(position)
+
+        @options = options
+      end
 
       sig { override.params(template: ::CSVPlusPlus::Template).void }
       # Write the +template+ to an Excel file
@@ -17,11 +26,16 @@ module CSVPlusPlus
       # @param template [Template] The template to write
       def write(template)
         ::CSVPlusPlus::Writer::RubyXLBuilder.new(
-          output_filename: ::T.must(@options.output_filename),
-          rows: template.rows,
+          options: @options,
           position: @position,
-          sheet_name: @options.sheet_name
+          rows: template.rows
         ).build_workbook.write(@options.output_filename)
+      end
+
+      sig { override.void }
+      # Write a backup of the current spreadsheet.
+      def write_backup
+        backup_file(@options)
       end
     end
   end