+++ spec/csv_plus_plus/entities/reference_spec.rb
@@ -2,7 +2,8 @@
 # frozen_string_literal: true
 
 describe ::CSVPlusPlus::Entities::Reference do
-  subject(:entity) { described_class.new(ref: 'rownum') }
+  let(:ref) { 'rownum' }
+  subject(:entity) { described_class.new(ref:) }
 
   describe '#==' do
     it { is_expected.to(eq(build(:reference, ref: 'rownum'))) }
@@ -16,6 +17,12 @@ describe ::CSVPlusPlus::Entities::Reference do
     subject { entity.evaluate(position) }
 
     it { is_expected.to(eq('rownum')) }
+
+    context 'with a cell reference' do
+      let(:ref) { 'C1' }
+
+      it { is_expected.to(eq('C1')) }
+    end
   end
 
   describe '#id' do