+++ docs/CHANGELOG.md
@@ -1,5 +1,3 @@
-## main/upcoming
-
 ## v0.2.0
 
 ### **Breaking Changes**
@@ -8,6 +6,8 @@
 
 ### Non-breaking Changes
 
+- Excel: fix the merging of existing values 
+- CSV: fix the merging of existing values 
 - Support merging in values from CSV (previously it would ignore/overwrite them)
 - Allow for more generous spacing in the csv section (and reflect this in the examples)
 - More type coverage